log_analyzer/requirements.txt
langchain-openai>=0.1.0
langgraph>=0.0.20
langchain-core>=0.2.0
python-dotenv>=1.0.0
langsmith>=0.1.0
